feat: use hostavatarsbutton, moved component to journeysadmin
