+++ apps/journeys-admin/src/components/JourneyQuickSettings/JourneyQuickSettings.tsx
@@ -1,17 +1,15 @@
 import { Box, Button, Stack, TextField, Typography } from '@mui/material'
 import { ReactElement } from 'react'
 
-import { noop } from 'lodash'
-import { ImageUpload } from '../../../../../../apps/journeys-admin/src/components/Editor/Slider/Settings/Drawer/ImageBlockEditor/CustomImage/ImageUpload'
+import { Chat } from '../Editor/Slider/Settings/CanvasDetails/Footer/Chat'
 
-import { Chat } from '../../../../../../apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/Chat'
-
-import { HostTitleFieldForm } from '../../../../../../apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/HostTab/HostForm/HostTitleFieldForm'
-import { PreviewItem } from '../../../../../../apps/journeys-admin/src/components/Editor/Toolbar/Items/PreviewItem'
+import { HostTitleFieldForm } from '../Editor/Slider/Settings/CanvasDetails/Footer/HostTab/HostForm/HostTitleFieldForm'
+import { PreviewItem } from '../Editor/Toolbar/Items/PreviewItem'
 
 import { ShareButton } from 'libs/journeys/ui/src/components/StepFooter/FooterButtonList/ShareButton'
+import { HostAvatarsButton } from '../Editor/Slider/Settings/CanvasDetails/Footer/HostTab/HostForm/HostAvatarsButton'
 
-export function JourneysQuickPage(): ReactElement {
+export function JourneyQuickSettings(): ReactElement {
   return (
     <Box sx={{ p: 4 }}>
       <Typography variant="h3"> Journeys Form Page</Typography>
@@ -21,7 +19,7 @@ export function JourneysQuickPage(): ReactElement {
       {/* Use imageblock on journey?  */}
       <Typography variant="h4">Your Profile Image: </Typography>
 
-      <ImageUpload selectedBlock={null} onChange={noop} />
+      <HostAvatarsButton />
       <Typography variant="h4"> Social media link: </Typography>
       <Box sx={{ p: 4 }}>
         <Chat />
