+++ apps/journeys-admin/src/components/JourneyQuickSettings/index.ts
@@ -1,0 +1,1 @@
+export { JourneysQuickPage } from './JourneyQuickSettings'
