--- libs/shared/ui/src/components/JourneysQuickPage/index.ts
@@ -1,1 +1,0 @@
-export { JourneysQuickPage } from './JourneysQuickPage'
