--- libs/shared/ui/src/components/JourneysShareableLinkPage/index.ts
@@ -1,1 +1,0 @@
-export { JourneysShareableLinkPage } from './JourneysShareableLinkPage'
